swarm/storage/mru: Documented resource.go extensively
moved some functions where they make most sense
